fix(gateway): authorize guarded session deletes
